+++ src/main/java/com/easypost/Constants.java
@@ -62,7 +62,7 @@ public abstract class Constants {
 
     public abstract static class CarrierAccountTypes {
         public static final List<String> CARRIER_TYPES_WITH_CUSTOM_WORKFLOW = ImmutableList.of("FedexAccount",
-                "UpsAccount", "FedexSmartpostAccount");
+                "FedexSmartpostAccount");
     }
 
     public abstract static class Http {