+++ src/test/java/com/easypost/fixtures/components/Claims.java
@@ -1,0 +1,10 @@
+package com.easypost.fixtures.components;
+
+import com.google.gson.annotations.SerializedName;
+
+import java.util.HashMap;
+
+public final class Claims {
+    @SerializedName ("basic")
+    public HashMap<String, Object> basic;
+}